pyproject.toml
[tool.poetry]
name = "hololivesubs"
version = "0.1.0"
description = ""
authors = ["Nachtalb <info@nachtalb.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ics = "^0.7.2"
yarl = "^1.9.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
bilibili-api-python = "^15.5.3"
tweepy = "^4.14.0"
rfeed = "^1.1.1"
pillow = "^10.0.0"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = { extras = ["black"], version = "^8.14.0" }


[tool.poetry.group.deve.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
ruff = "^0.0.289"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
preview = true
target-version = ["py311"]

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.11"
show_error_codes = true
pretty = true
strict = true
